Replace some files and fix bugs
